requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.1.0
auth0-python==4.7.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.8
Django==4.2.14
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
git-filter-repo==2.45.0
idna==3.7
kombu==5.3.7
msgpack==1.1.0
multidict==6.0.5
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
razorpay==1.4.2
redis==5.0.8
requests==2.32.3
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
twilio==9.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
